GooglePlay and Amazon scores now support 64bit int values.

+++ Assets/Plugins/Reign/Android/Amazon_GameCircle_ScorePlugin.cs
@@ -82,7 +82,7 @@ namespace Reign.Plugin
 			native.CallStatic("ReportAchievement", findAchievementID(achievementID), (percentComplete / found.PercentCompletedAtValue) * 100f);
 		}
 
-		public void ReportScore (string leaderboardID, int score, ReportScoreCallbackMethod callback, MonoBehaviour services)
+		public void ReportScore (string leaderboardID, long score, ReportScoreCallbackMethod callback, MonoBehaviour services)
 		{
 			reportScoreCallback = callback;
 			native.CallStatic("ReportScore", findLeaderboardID(leaderboardID), score);
